brewyourtech/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
CurrencyConverter==0.18.1
Django==5.1.4
django-bootstrap5==24.3
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
sqlparse==0.5.3
urllib3==2.2.3
wheel==0.44.0